// src/interface/erpProtheus/invoice/invoice.interface.ts
export interface IInvoiceErpProtheus {
	branchID: string;
	orderERP: string;
	orderCRM: string;
	invoice: string;
	series: string;
	customer: string;
	store: string;
	dateIssue: string;
	hourIssue: string;
	dateAuthorization: string;
	hourAuthorization: string;
	key: string;
	volume: string;
	DANFE: any;
	XML: any;
}